Update 4/13
Configurator is now working step for step, there's the left side (information/display) in relation to the current , and the right side (form) which is the necessary input needed from user. 

There is also the custom Radio Selection on the 2nd tab, which allows for custom images to be selected instead of radio buttons with text. The css for the custom radio selection is inside css/custom/radio.css .